Parcel entity is ready entirely!

+++ dotNet5782_9627_5223/PL/HelpingClasses/Validation/ValidationRules.cs
@@ -38,7 +38,7 @@ namespace PL
         public override ValidationResult Validate(object value, CultureInfo cultureInfo)
         {
             if (value == null) return new ValidationResult(false, "Required.");
-            return int.Parse(value.ToString()) > 0 ? ValidationResult.ValidResult : new ValidationResult(false, "Input has to contain a real positive value.");
+            return long.Parse(value.ToString()) > 0 ? ValidationResult.ValidResult : new ValidationResult(false, "Input has to contain a real positive value.");
         }
     }
 
@@ -47,7 +47,7 @@ namespace PL
         public override ValidationResult Validate(object value, CultureInfo cultureInfo)
         {
             if (value == null) return new ValidationResult(false, "Required.");
-            return int.Parse(value.ToString()) >= 0 ? ValidationResult.ValidResult : new ValidationResult(false, "Input has to contain a positive value.");
+            return long.Parse(value.ToString()) >= 0 ? ValidationResult.ValidResult : new ValidationResult(false, "Input has to contain a positive value.");
         }
     }
 
@@ -90,9 +90,10 @@ namespace PL
     {
         public override ValidationResult Validate(object value, CultureInfo cultureInfo)
         {
-            return int.Parse(value.ToString()) != 0 ? ValidationResult.ValidResult : new ValidationResult(false, "Not initalized field.");
+            return long.Parse(value.ToString()) != 0 ? ValidationResult.ValidResult : new ValidationResult(false, "Not initalized field.");
         }
     }
+
 }
